+++ example/lib/main.dart
@@ -44,10 +44,9 @@ class _MyAppState extends State<MyApp> {
     // must initialize the FileDownloader by passing that alternative storage
     // object on the first call to FileDownloader.
     // As an example, this example app has implemented a backing using
-    // the sqflite package (works for Android/iOS only and isn't production
-    // ready -> only use as an example).
-    // To try that, uncomment the following line, which
-    // will initialize the downloader with that storage solution.
+    // the sqflite package (works for Android/iOS only).
+    // To try that the default version, comment the following line, which
+    // will initialize the downloader with teh default storage solution.
     FileDownloader(persistentStorage: SqlitePersistentStorage());
 
     // Configure the downloader by registering a callback and configuring