+++ example/lib/main.dart
@@ -45,9 +45,9 @@ class _MyAppState extends State<MyApp> {
     // object on the first call to FileDownloader.
     // As an example, this example app has implemented a backing using
     // the sqflite package (works for Android/iOS only).
-    // To try that the default version, comment the following line, which
-    // will initialize the downloader with teh default storage solution.
-    FileDownloader(persistentStorage: SqlitePersistentStorage());
+    // To try that SQLite version, uncomment the following line, which
+    // will initialize the downloader with the SQLite storage solution.
+    // FileDownloader(persistentStorage: SqlitePersistentStorage());
 
     // Configure the downloader by registering a callback and configuring
     // notifications